Move Datasource\LocatorInterface to Datasource\Loctor\LocatorInterface
