+++ src/Datasource/Locator/LocatorInterface.php
@@ -14,7 +14,9 @@ declare(strict_types=1);
  * @since         4.1.0
  * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Datasource;
+namespace Cake\Datasource\Locator;
+
+use Cake\Datasource\RepositoryInterface;
 
 /**
  * Registries for repository objects should implement this interface.
