+++ src/Datasource/ModelAwareTrait.php
@@ -17,6 +17,7 @@ declare(strict_types=1);
 namespace Cake\Datasource;
 
 use Cake\Datasource\Exception\MissingModelException;
+use Cake\Datasource\Locator\LocatorInterface;
 use InvalidArgumentException;
 use UnexpectedValueException;
 
@@ -47,7 +48,7 @@ trait ModelAwareTrait
     /**
      * A list of overridden model factory functions.
      *
-     * @var (callable|\Cake\Datasource\LocatorInterface)[]
+     * @var (callable|\Cake\Datasource\Locator\LocatorInterface)[]
      */
     protected $_modelFactories = [];
 
@@ -134,14 +135,14 @@ trait ModelAwareTrait
      * Override a existing callable to generate repositories of a given type.
      *
      * @param string $type The name of the repository type the factory function is for.
-     * @param callable|\Cake\Datasource\LocatorInterface $factory The factory function used to create instances.
+     * @param callable|\Cake\Datasource\Locator\LocatorInterface $factory The factory function used to create instances.
      * @return void
      */
     public function modelFactory(string $type, $factory): void
     {
         if (!$factory instanceof LocatorInterface && !is_callable($factory)) {
             throw new InvalidArgumentException(sprintf(
-                '`$factory` must be an instance of Cake\Datasource\LocatorInterface or a callable.'
+                '`$factory` must be an instance of Cake\Datasource\Locator\LocatorInterface or a callable.'
                 . ' Got type `%s` instead.',
                 getTypeName($factory)
             ));
